+++ packages/wds-engine/src/index.ts
@@ -1,5 +1,3 @@
-export { css, CacheProvider, ClassNames, keyframes } from '@emotion/react';
-export { default as createCache } from '@emotion/cache';
 export {
   theme,
   lightOriginTheme,
@@ -12,3 +10,4 @@ export * from './components';
 export * from './context';
 export * from './hooks';
 export * from './types';
+export * from './utils';
