+++ packages/wds-engine/src/utils/emotion.ts
@@ -1,0 +1,3 @@
+'use client';
+export { css, CacheProvider, ClassNames, keyframes } from '@emotion/react';
+export { default as createCache } from '@emotion/cache';
