+++ packages/wds-engine/src/utils/interpolation.ts
@@ -1,0 +1,19 @@
+import type { Interpolation } from '@emotion/react';
+import type { Theme } from '@wanteddev/wds-theme';
+
+export const interpolationTheme = (
+  expressions: Interpolation<Theme>,
+  theme: Theme,
+): Interpolation<Theme> => {
+  if (Array.isArray(expressions)) {
+    return expressions.map((expression) =>
+      interpolationTheme(expression, theme),
+    );
+  }
+
+  if (typeof expressions === 'function') {
+    return expressions(theme);
+  }
+
+  return expressions;
+};
