+++ packages/wds-engine/tsup.config.ts
@@ -9,11 +9,11 @@ export default defineConfig(
     dts: 'src/index.ts',
     onSuccess: () =>
       injectUseClient([
-        './dist/box/**/index.{js,mjs}',
-        './dist/theme-provider/**/index.{js,mjs}',
+        './dist/components/**/index.{js,mjs}',
         './dist/context/**/index.{js,mjs}',
-        './dist/hooks/*/*.{js,mjs}',
+        './dist/hooks/**/*.{js,mjs}',
         './dist/global/**/index.{js,mjs}',
+        './dist/utils/**/emotion.{js,mjs}',
       ]),
   }),
 );
